requirements.txt
streamlit>=1.26.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
yt-dlp>=2023.10.13
python-dotenv>=1.0.0
scikit-learn>=1.3.0
gensim>=4.3.1

# Firebase dependencies
firebase-admin>=6.2.0
pyrebase4>=4.7.1
